+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Fixed
 
 - Fix output format of `slumber request --dry-run ...` to match `slumber request --verbose`
+- Fix `curl` output for URL-encoded and multipart forms
 
 ## [3.0.1] - 2025-02-19